Refactoriza estadisticas sin queries explicitas
- Elimina @Query de los repositorios de prestamos y sanciones.

- Mueve el agrupamiento y ordenamiento al servicio de estadisticas.

- Agrega LibroRepository para resolver datos de libros con JPA simple.

- Cambia los totales de long a Integer en los DTOs.

+++ ms-prestamos/src/main/java/com/scrip/msprestamos/dto/LibroMasPrestadoResponse.java
@@ -6,6 +6,6 @@ public record LibroMasPrestadoResponse(
         UUID libroId,
         String titulo,
         String isbn,
-        long totalPrestamos
+        Integer totalPrestamos
 ) {
 }
