+++ ms-prestamos/src/main/java/com/scrip/msprestamos/dto/PrestamosPeriodoResponse.java
@@ -5,6 +5,6 @@ import java.time.LocalDate;
 public record PrestamosPeriodoResponse(
         LocalDate fechaInicio,
         LocalDate fechaFin,
-        long totalPrestamos
+        Integer totalPrestamos
 ) {
 }
